emitter post

+++ integration/Context/Request/EmitterRequest.cs
@@ -5,7 +5,7 @@ public class EmitterRequest
     public int? idAsuPro { get; set; }
     public int? idConsumer { get; set; }
     public int? idConsumerType { get; set; }
-    public int? amount { get; set; }
+    public float? amount { get; set; }
     public string? consumerAddress { get; set; }
     public string? accountingType { get; set; } //норматив
     public string? contractNumber { get; set; }
